Extract test helpers so that LoggedTest and general helpers are generally usable, but Webview stuff is marked appropriately
