+++ test/test_flow.rb
@@ -2,7 +2,7 @@
 
 require "test_helper"
 
-class TestWebviewFlow < ScarpeTest
+class TestWebviewFlow < ScarpeWebviewTest
   def test_it_accepts_a_height
     flow = Scarpe::WebviewFlow.new("width" => 7, "height" => 25, "margin" => 4, "padding" => 5, "shoes_linkable_id" => 1) do
       "fishes that flop"
