+++ test/test_html.rb
@@ -2,7 +2,7 @@
 
 require "test_helper"
 
-class TestHTML < ScarpeTest
+class TestHTML < ScarpeWebviewTest
   def test_works_without_content
     subject = ::Scarpe::HTML.render(&:div)
